+++ Task_1_1_2/src/main/java/ru/nsu/borodkin/blackjack/Card.java
@@ -1,4 +1,4 @@
-package ru.nsu.borodkin;
+package ru.nsu.borodkin.blackjack;
 
 /**
  * Класс, представляющий карту в игре Blackjack.
