+++ Task_1_1_2/src/main/java/ru/nsu/borodkin/blackjack/Deck.java
@@ -1,4 +1,4 @@
-package ru.nsu.borodkin;
+package ru.nsu.borodkin.blackjack;
 
 import java.util.Collections;
 import java.util.LinkedList;
