+++ Task_1_1_2/src/main/java/ru/nsu/borodkin/blackjack/Player.java
@@ -1,4 +1,4 @@
-package ru.nsu.borodkin;
+package ru.nsu.borodkin.blackjack;
 
 import java.util.ArrayList;
 import java.util.List;
