+++ Task_1_1_2/src/test/java/ru/nsu/borodkin/AppTest.java
@@ -4,6 +4,7 @@ import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 
 import org.junit.jupiter.api.Test;
+import ru.nsu.borodkin.blackjack.BlackjackGame;
 
 import static org.junit.jupiter.api.Assertions.*;
