+++ Task_1_1_2/src/test/java/ru/nsu/borodkin/blackjack/DeckTest.java
@@ -1,6 +1,8 @@
-package ru.nsu.borodkin;
+package ru.nsu.borodkin.blackjack;
 
 import org.junit.jupiter.api.Test;
+import ru.nsu.borodkin.blackjack.Card;
+import ru.nsu.borodkin.blackjack.Deck;
 
 import java.util.ArrayList;
 import java.util.List;
